Enable Windows native support for ANSI escape sequences
This was brought to my attention in the coloredlogs issue tracker:

- https://github.com/xolox/python-coloredlogs/issues/71
- https://github.com/xolox/python-coloredlogs/pull/72

My experiences with ANSI escape sequences started out as part of the
coloredlogs package but eventually I moved the support for ANSI escape
sequences to my humanfriendly package. This explains how it now makes
sense to integrate the Windows native ANSI escape sequence support in
the humanfriendly package as well.