Stop hiding text cursor when spinner is shown
While the code to hide the text cursor when a spinner is shown seems
simple enough, I've had it happen to me on numerous occasions that the
text cursor is successfully hidden but visibility is never restored!

I've now decided that I don't care to investigate what's going on here,
instead I will simply remove the feature, given that it's such a small
detail in the grand scheme of things (certainly not worth the bug).